requirements.txt
boto3~=1.28.0
streamlit~=1.50.0
pandas~=2.3.2
python-pptx~=1.0.2
openpyxl
requests

botocore~=1.31.85